Move GeoJSON feature conversion logic to its own module
